+++ projects/ids-enterprise-ng/src/lib/about/soho-about.ref.ts
@@ -1,7 +1,9 @@
+import { NgZone } from "@angular/core";
 /**
  * Wrapper for the jQuery about control.
  */
 export class SohoAboutRef {
+  private eventGuard: SohoAboutDialogEventGuard | null | undefined = {};
 
   /** Selector referencing the about-dialog after it has been moved to the dialog container. */
   private jQueryElement?: JQuery;
@@ -114,7 +116,7 @@ export class SohoAboutRef {
    *
    * @param placeholder for the about dialog; defaults to the body.
    */
-  constructor() {
+  constructor(private ngZone: NgZone) {
     this._placeholder = $('body');
   }
 
@@ -128,10 +130,12 @@ export class SohoAboutRef {
     this._placeholder.about(this._options);
 
     // Need a better way of locating the about dialog dialog.
-    this.jQueryElement = this._placeholder.find('about');
-
-    // Get the api.
+    this.jQueryElement = this._placeholder.find('.about');
     this._about = this.jQueryElement.data('about');
+
+    this.jQueryElement?.off('beforeopen.about').on('beforeopen.about', ((event: any) => this.ngZone.run(() => this.onBeforeOpen(event))));
+    this.jQueryElement?.off('close.about').on('close.about', ((event: any) => this.ngZone.run(() => this.onClose(event))));
+    this.jQueryElement?.off('afterclose.about').on('afterclose.about', ((event: any) => this.ngZone.run(() => this.onAfterClose(event))));
 
     return this;
   }
@@ -148,4 +152,99 @@ export class SohoAboutRef {
     return this;
   }
 
+  /**
+   * Registers a before open guard.
+   *
+   * @param eventFn - the function to call before openning the dialog.
+   */
+  beforeOpen(eventFn: (dialogRef?: SohoAboutRef) => boolean): SohoAboutRef {
+    (this.eventGuard as any).beforeOpen = eventFn;
+    return this;
+  }
+
+  /**
+   * Registers a closed guard.
+   *
+   * @param eventFn - the function to call before openning the dialog.
+   */
+  closed(eventFn: (dialogRef?: SohoAboutRef) => boolean): SohoAboutRef {
+    (this.eventGuard as any).closed = eventFn;
+    return this;
+  }
+
+  /**
+   * Registers a after close guard.
+   *
+   * @param eventFn - the function to call before openning the dialog.
+   */
+  afterClose(eventFn: (dialogRef?: SohoAboutRef) => boolean): SohoAboutRef {
+    (this.eventGuard as any).afterClose = eventFn;
+    return this;
+  }
+
+  /**
+   * Handles the 'beforeOpen' event, fired before the modal dialog
+   * has been opened.
+   *
+   * @param event - full event object.
+   *
+   * @return true if the dialog can be opened; otherwise false if veoted.
+   */
+  private onBeforeOpen(_event: any): boolean {
+    const fn: Function = (this.eventGuard?.beforeOpen as any);
+    return fn ? fn.call(this.eventGuard, this) : true;
+  }
+
+  /**
+   * Handles the 'closed' event, fired before the modal dialog
+   * has been opened.
+   *
+   * @param event - full event object.
+   *
+   * @return true if the dialog can be opened; otherwise false if veoted.
+   */
+  private onClose(_event: any): boolean {
+    const fn: Function = (this.eventGuard?.closed as any);
+    return fn ? fn.call(this.eventGuard, this) : true;
+  }
+
+  /**
+   * Handles the 'afterClose' event, fired before the modal dialog
+   * has been opened.
+   *
+   * @param event - full event object.
+   *
+   * @return true if the dialog can be opened; otherwise false if veoted.
+   */
+  private onAfterClose(_event: any): boolean {
+    const fn: Function = (this.eventGuard?.afterClose as any);
+    return fn ? fn.call(this.eventGuard, this) : true;
+  }
+}
+
+
+/**
+ * Event Handlers.
+ */
+export interface SohoAboutDialogEventGuard {
+  /**
+   * Invoked before a modal is opened.
+   *
+   * @return false to veto the open action; otherwise true.
+   */
+  beforeOpen?(dialogRef: SohoAboutRef): boolean;
+
+  /**
+   * Invoked when closing a modal.
+   *
+   * @return false to veto the open action; otherwise true.
+   */
+  closed?(dialogRef: SohoAboutRef): boolean;
+
+  /**
+   * Invoked after the modal is closed.
+   *
+   * @return false to veto the open action; otherwise true.
+   */
+  afterClose?(dialogRef: SohoAboutRef): boolean;
 }